+++ saleor/app/manifest_validations.py
@@ -9,6 +9,7 @@ from semantic_version import NpmSpec, Version
 from semantic_version.base import Range
 
 from .. import __version__
+from ..giftcard.const import GIFT_CARD_PAYMENT_GATEWAY_ID
 from ..graphql.core.utils import str_to_enum
 from ..graphql.webhook.subscription_query import SubscriptionQuery
 from ..permission.enums import (
@@ -149,6 +150,15 @@ def clean_manifest_data(manifest_data, raise_for_saleor_version=False):
                 code=AppErrorCode.UNIQUE.value,
             )
         )
+
+    if manifest_data.get("id") == GIFT_CARD_PAYMENT_GATEWAY_ID:
+        errors["identifier"].append(
+            ValidationError(
+                "App with this identifier cannot be installed",
+                code=AppErrorCode.UNIQUE.value,
+            )
+        )
+
     if not errors:
         _clean_extensions(manifest_data, app_permissions, errors)
         _clean_webhooks(manifest_data, errors)